+++ lib/trackdown/ip_locator.rb
@@ -4,6 +4,7 @@ require_relative 'location_result'
 require_relative 'ip_validator'
 require_relative 'providers/auto_provider'
 require_relative 'providers/cloudflare_provider'
+require_relative 'providers/cloudfront_provider'
 require_relative 'providers/maxmind_provider'
 
 module Trackdown
@@ -11,7 +12,7 @@ module Trackdown
     class << self
       # Locate an IP address using the configured provider
       # @param ip [String] The IP address to locate
-      # @param request [ActionDispatch::Request, nil] Optional Rails request object for Cloudflare provider
+      # @param request [#env, nil] Optional Rack-compatible request object for CDN providers
       # @return [LocationResult] The location information
       def locate(ip, request: nil)
         IpValidator.validate!(ip)
@@ -32,6 +33,8 @@ module Trackdown
           Providers::AutoProvider
         when :cloudflare
           Providers::CloudflareProvider
+        when :cloudfront
+          Providers::CloudfrontProvider
         when :maxmind
           Providers::MaxmindProvider
         else